refactor: simplify imports with index in Infra
