+++ src/infra/cryptography/bcrypt-adapter.ts
@@ -1,6 +1,5 @@
 import bcrypt from 'bcrypt'
-import { type Hasher } from '@/application/protocols/cryptography/hasher'
-import { type HashComparer } from '@/application/protocols/cryptography/hash-comparer'
+import { type Hasher, type HashComparer } from '@/application/protocols/cryptography'
 
 export class BcryptAdapter implements Hasher, HashComparer {
   constructor(private readonly salt: number) {}
