+++ src/infra/db/mongodb/dynamic/authentication/index.ts
@@ -1,0 +1,1 @@
+export * from './dynamic-account-mongo-repository'
