+++ src/infra/db/mongodb/static/authentication/index.ts
@@ -1,0 +1,1 @@
+export * from './static-account-mongo-repository'
