+++ src/infra/validators/email-validator-adapter.ts
@@ -1,5 +1,5 @@
 import validator from 'validator'
-import { type EmailValidator } from '@/validation/protocols/email-validator'
+import { type EmailValidator } from '@/validation/protocols'
 
 export class EmailValidatorAdapter implements EmailValidator {
   public isValid = (email: string): boolean => {
